get_top_web3_games/requirements.txt
airplanesdk==0.3.22
python-dotenv==1.0.0
selenium==4.8.3
